Move S3DataStore to separate module so boto3 isn't loaded for file things
