+++ switcheroo/data_store/__init__.py
@@ -1,0 +1,1 @@
+from switcheroo.data_store.data_stores import DataStore, FileSystemDataStore
